requirements.txt
Flask==2.3.2
Flask_Cors==4.0.0
Flask_HTTPAuth==4.8.0
flask_restx==1.3.0
langchain==0.3.14
langchain-community==0.3.14
Requests>=2.28.0,<2.33.0
weaviate-client<4.0.0
Werkzeug==3.1.3
tiktoken==0.8.0
pytesseract
pillow
pdf2image
tqdm
anthropic
jieba
